[Lab1] Implement some part of class `HeapFile` (#14)
* [Lab1] Implement some part of class `HeapFile`

* [Lab1] Fix for `RecordId`

+++ src/java/simpledb/HeapPageId.java
@@ -45,7 +45,7 @@ public class HeapPageId implements PageId {
    * @see BufferPool
    */
   public int hashCode() {
-    return Integer.parseInt(Integer.toString(tid) + pgNo);
+    return ((int) Long.parseLong(Integer.toString(tid) + pgNo));
   }
 
   /**
